brought registry methods and RegistrySupplier back together to register mod objects in proper sequence
