+++ common/src/main/java/net/phoboss/decobeacon/DecoBeacon.java
@@ -5,10 +5,10 @@ import dev.architectury.utils.Env;
 import dev.architectury.utils.EnvExecutor;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.phoboss.decobeacon.blocks.ModBlockEntitiesRegistry;
-import net.phoboss.decobeacon.blocks.ModBlocksRegistry;
-import net.phoboss.decobeacon.client.rendering.ModRenderingRegistry;
-import net.phoboss.decobeacon.items.ModItemsRegistry;
+import net.phoboss.decobeacon.blocks.ModBlockEntities;
+import net.phoboss.decobeacon.blocks.ModBlocks;
+import net.phoboss.decobeacon.client.rendering.ModRendering;
+import net.phoboss.decobeacon.items.ModItems;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,9 +17,9 @@ public class DecoBeacon {
     public static final String MOD_ID = "decobeacon";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
     public static void init() {
-        ModBlocksRegistry.registerAll();
-        ModBlockEntitiesRegistry.registerAll();
-        ModItemsRegistry.registerAll();
+        ModBlocks.registerAll();
+        ModBlockEntities.registerAll();
+        ModItems.registerAll();
         EnvExecutor.runInEnv(Env.CLIENT, () -> DecoBeacon.Client::initClient);
     }
 
@@ -30,7 +30,7 @@ public class DecoBeacon {
             ClientLifecycleEvent.CLIENT_STARTED.register((client) ->
                 {
                     LOGGER.info("Client starting!");
-                    ModRenderingRegistry.registerAll();
+                    ModRendering.registerAll();
                 });
             ClientLifecycleEvent.CLIENT_STOPPING.register((client) ->
                 {
