+++ common/src/main/java/net/phoboss/decobeacon/client/rendering/ModRendering.java
@@ -7,7 +7,7 @@ import net.phoboss.decobeacon.blocks.ModBlockEntities;
 import net.phoboss.decobeacon.blocks.ModBlocks;
 import net.phoboss.decobeacon.blocks.decobeacon.DecoBeaconBlockEntityRenderer;
 
-public class ModRenderingRegistry {
+public class ModRendering {
 
     public static void registerRenderType() {
         RenderTypeRegistry.register(RenderLayer.getTranslucent(),ModBlocks.DECO_BEACON.get());
